A3: test intersection distance. Found bug
I thought that intersection of two wires is a common point between these
two wires. I was right, but I forgot that a wire has an infinite amount
of points in it -- as every curve.

Because of that I found intersection of wires as two common *angle*
points between two wires, which is wrong.

Now I should develop another algorithm of finding the intersection.