A3: second star!
It was soooo easy given the data structure for Wire I created.
As all points in a wire lie at an integer grid and all points in a wire
are sorted according to their position at the wire, length to the point is
equal to it's position in a wire vec!

This insight was such a great thing! I love it!